Support routes over construction railways
Sometimes railways are under construction, nevertheless a route is set over it.

That should be ignored, as it is unlogic to have a route over a railway being in construction.

Nevertheless that should not be a problem for the tool. Ignoring it will leed to strange results (no warning, the route is shorter than in reality, the signals on the construction track are not shown).