Idea of fail safe state determination
This is my idea how a fail safe determination of states can be done, including throwing warnings.

Please not that I did not implement a Zs 3v check yet. That would be needed: If Zs 3v is not possible, show only Ks2! That is the real behavoiur: If the Zs 3v is broken/defect, Ks2 is shown. If it is not tagged in OSM database, that behaviour should be emulated.